Adds a base `ts_config()` rule.
This is intended to be the default for all TypeScript compilations in the workspace. It should either be used directly or extended with required modifications on top of it.

I generated the `tsconfig.json` itself with `npx typescript --init`. I updated a few of the fields to enable sourcemaps and be a little more strict. I also configured the basic compilation options according to this comment, as Bazel appears to load Node v12.13.0.

I needed to install and add `@types/node` to the renderer in order for `console` to be available.

I'm debating making a `ts_project()` macro which automatically adds the `//:tsconfig` label and `source_map` flags. For now, I'll leave it is as, but we may want to update to that in the future.