Adds `--template` option to choose which template to render.
I opted to use `yargs` for option handling because I'm a helpless sheeple who can't make my own decisions.

Unfortunately, Nunjucks seems to expect all its templates to be in the same directory, so we can't just give it a file path and expect it to render. Instead, we need to set it up with a particular directory and then load the file by name in that directory. It's really awkward and will likely be a pain in the future.