Adds simple CI.
This merely runs `npm test`, which in turn, runs `bazel test //...`. This will execute all tests in the workspace.

Note that this will **not** build targets that aren't depended upon by a test. If a target should be verified to build by CI, then a `build_test()` rule can be used.

https://github.com/bazelbuild/bazel-skylib/blob/master/rules/build_test.bzl