Adds `njk_template()` BUILD rule.
For now, all this does is wrap the renderer binary since dependencies are not tracked or implemented as of yet.

The test for this is simply that the build of an example was successful. However we should really compare against a golden value. Unfortunately, I can't find a good way of doing that without a wrapper shell script which requires the Bash runfiles system. For the life of me I cannot get Bash runfiles to work, so I'm just deferring a goldens test for now.