Adds stricter flags to Nunjucks renderer and updates error DX.
With these new flags, most importantly a template will fail if a user references a symbol that is not known to the renderer.

This also updates the error flow to give concise error messages from Nunjucks but still provide full stack traces for internal errors.

An additional test is added to verify the error use case.