Fixes VSCode TypeScript intellisense.
I noticed that VSCode intellisense was not using settings from the tsconfig.json file. I made two changes to get this working.

First, I updated VSCode to use the same version of TypeScript as was installed in the project. This keeps versions consistent.

Second, I excluded `bazel-*/` from the tsconfig. The TypeScript language service appeared to be processing files outside the project itself and either failing or hanging as a result.

With both of these fixes, VSCode appears to give errors in line with `bazel` executions.