Adds basic dependency management.
This introduces a `deps` attribute to `njk_template()` to allow templates to depend on each other.

There's a lot going on in this commit, but I'm being a little lazy and not splitting it up properly.

This essentially introduces a new loader which is responsible for reading files from Bazel to be used in a Nunjucks template. The files loaded are provided via flags which generate the list from transitive dependencies.

There are two major problems not addressed in this commit:
1. Strict dependencies. Nothing enforces that a template only imports files in its `deps`, it is possible to import a transitive dependency.
2. Global data. All the templates used in a Nunjucks render share the same input data, there is no built-in manner of making a `foo` variable available **only** to `library.njk` and **not** to `page.njk`.

I also changed from `ts_project()` to `ts_library()` because I couldn't figure out how to compose the former. The `deps` attribute just doesn't seem to work the way I think it does and I couldn't find any good examples doing what I wanted. It was just easier to switch to `ts_library()` which conformed a bit better to my mental model.